chore: 质量门禁落地——fmt 统一、clippy 清零、workspace lints
阶段一+二（渐进式质量门禁）：
- cargo fmt：全仓 37 个文件统一格式
- cargo clippy --fix + 手工修复：33 条告警清零
  （unit let-binding ×15、redundant closure ×8、&PathBuf→&Path、
   needless clone on Copy 等）
- [workspace.lints]：clippy all = deny、rust unsafe_code = warn，
  全部 crate 经 [lints] workspace = true 继承
- mo-ui 因 objc 宏 check-cfg 需本地 lints 段，手动复制 workspace 策略
- unsafe 收敛：icon.rs 为全仓唯一 FFI 层，模块级 allow + SAFETY 注释
- 验证：clippy --all-targets 零告警、fmt --check 干净、87 测试全绿

+++ crates/mo-fs/src/lib.rs
@@ -72,7 +72,7 @@ pub(crate) fn to_dir_error(e: std::io::Error) -> MoError {
 pub(crate) fn file_id_for(path: &Path) -> FileId {
     if let Ok(meta) = std::fs::metadata(path) {
         use std::os::unix::fs::MetadataExt;
-        return FileId::new(meta.dev() as u64, meta.ino() as u128);
+        return FileId::new(meta.dev(), meta.ino() as u128);
     }
     FileId::synthetic(path)
 }
